pyproject.toml
[project]
name = "sathya-project"
version = "0.1.0"
description = "Stripe MRR Dashboard Pipeline for Digital Signage SaaS"
requires-python = ">=3.12"
dependencies = [
    "stripe>=8.0.0",
    "python-dotenv>=1.0.0",
    "google-cloud-bigquery>=3.20.0",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
]
